+++ src/helpers/convert-glob.ts
@@ -1,6 +1,6 @@
 import globToRegexp from 'glob-to-regexp';
 
 export default function parseGitGlob(path: string): RegExp {
-	const globPath = `**/${path}`.replace(/\[:(space|digit):\]/, (_, val) => `\\${val[0]}`);
+	const globPath = `**/${path}`.replace(/\\/g, '/').replace(/\[:(space|digit):\]/g, (_, val) => `\\${val[0]}`);
 	return globToRegexp(globPath, { globstar: true, extended: true });
 }