+++ changelog.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## Next
+- Fixed git config entries containing backslashes causing a crash.
+
 ## 2.1.4
 *2021-09-29*
 - Fixed gitignore entries containing dots not being parsed properly.